separate test file and benchmark_test file
