+++ finder_test.go
@@ -741,195 +741,3 @@ func TestNewFinderKeys(t *testing.T) {
 		})
 	}
 }
-
-// benchmark tests
-
-func BenchmarkNewFinder_Val(b *testing.B) {
-	var f *Finder
-	var e error
-
-	testStructVal := newFinderTestStruct()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		f, e = NewFinder(testStructVal)
-		if e == nil {
-			_ = f
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", e)
-		}
-	}
-}
-
-func BenchmarkNewFinder_Ptr(b *testing.B) {
-	var f *Finder
-	var e error
-
-	testStructPtr := newFinderTestStructPtr()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		f, e = NewFinder(testStructPtr)
-		if e == nil {
-			_ = f
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", e)
-		}
-	}
-}
-
-func BenchmarkToMap_1FindOnly(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Find("String").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkToMap_2FindOnly(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Find("String", "Int64").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkToMap_1Struct_1Find(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Into("FinderTestStruct2").Find("String").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkToMap_1Struct_1Find_2Pair(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Into("FinderTestStruct2").Find("String").Into("FinderTestStruct2Ptr").Find("String").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkToMap_2Struct_1Find(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Into("FinderTestStruct2", "FinderTestStruct3").Find("String").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkToMap_2Struct_2Find(b *testing.B) {
-	var m map[string]interface{}
-
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		m, err = f.Into("FinderTestStruct2", "FinderTestStruct3").Find("String", "Int").ToMap()
-		if err == nil {
-			_ = m
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkNewFinderKeys_yml(b *testing.B) {
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		fks, err := NewFinderKeys("examples/finder_from_conf", "ex_test1_yml")
-		if err == nil {
-			_ = f.FromKeys(fks)
-			f.Reset()
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
-
-func BenchmarkNewFinderKeys_json(b *testing.B) {
-	f, err := NewFinder(newFinderTestStructPtr())
-	if err != nil {
-		b.Fatalf("NewFinder() occurs unexpected error: %v", err)
-		return
-	}
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		fks, err := NewFinderKeys("examples/finder_from_conf", "ex_test1_json")
-		if err == nil {
-			_ = f.FromKeys(fks)
-			f.Reset()
-		} else {
-			b.Fatalf("abort benchmark because error %v occurd.", err)
-		}
-	}
-}
